refactor(auth): extract FixedClockConfig to top-level, inline request post-processor (SPP-83)
Apply project standards to SecurityFilterChainIT:
- @TestConfiguration moved out of inner static class into a package-private
  top-level FixedClockConfig.java (no-inner-classes rule).
- private RequestPostProcessor remoteAddr(...) helper inlined at call sites
  (TESTING_STANDARDS §5 Pattern 6 — no private fixture helpers in test
  classes; for one-off lambdas inline rather than building a fixture).

No behavior change; ./gradlew :apps:auth:main:integrationTest still green.
